Se agregan views para login y logout

+++ perritos_app/views.py
@@ -2,14 +2,18 @@ from django.forms import BaseModelForm
 from django.shortcuts import render
 from django.http import HttpResponse, HttpResponseRedirect
 from django.views.generic.edit import FormView, UpdateView
+from django.views.generic import TemplateView
 from django.urls import reverse_lazy
 from django.contrib import messages
+from django.contrib.auth.mixins import UserPassesTestMixin, LoginRequiredMixin
 
 from rest_framework import viewsets
+from rest_framework.views import APIView
 from rest_framework import status
 from rest_framework import generics
 from rest_framework.response import Response
 from rest_framework.renderers import TemplateHTMLRenderer, JSONRenderer, BrowsableAPIRenderer
+from drf_spectacular.utils import extend_schema
 
 from .serializers import PerroSerizalizer
 from .models import Perro, PerroFotos
@@ -18,39 +22,60 @@ from .forms import PerrosForm, PerroFotosForm
 from .permissions import IsAdminOrReadOnly
 
 # Create your views here.
-def index(request):
-    return render(request, 'index.html')
+class IndexView(TemplateView):
+    template_name='index.html'
+    #permission_classes = [IsAdminOrReadOnly]
+    #renderer_classes = [TemplateHTMLRenderer]
+    #serializer_class = None 
+
+    #def get(self, request):
+    #    return Response(template_name='index.html')
 
 # View para listar todas las entradas o solo de a una
 class PerritosViewSet(viewsets.ModelViewSet):
-    serializer_class = PerroSerizalizer
     permission_classes = [IsAdminOrReadOnly]
-    renderer_classes = [TemplateHTMLRenderer, BrowsableAPIRenderer, JSONRenderer]
+    queryset = Perro.objects.all().filter(status=True).order_by('id')
+    serializer_class = PerroSerizalizer
+    renderer_classes = [TemplateHTMLRenderer, JSONRenderer]
+    template_name='perritos_list.html'
 
-    def get_queryset(self):
-        return Perro.objects.all().filter(status=True)
-
+    def render_response(self, data, request, template_name=None, context=None):
+        """Función para manejar la respuesta HTML o JSON."""
+        context = context or {}
+        if request.accepted_renderer.format == 'html':
+            return Response({**context, 'perros': data}, template_name=template_name)
+        return Response(data)
+    
     def list(self, request, *args, **kwargs):
         queryset = self.get_queryset()
-        if request.accepted_renderer.format == 'html':
-            return Response({'perros': queryset}, template_name='perritos_list.html')
-        serializer = self.get_serializer(queryset, many=True)
-        return Response(serializer.data)
+        page = self.paginate_queryset(queryset)
+        data = self.get_serializer(page or queryset, many=True).data
+        
+        if page is not None:
+            paginated_response = self.get_paginated_response(data)
+            pagination_data = {
+                'total_pages': paginated_response.data.get('count') // self.paginator.page_size + 1,
+                'current_page': self.paginator.page.number,
+                'has_next': self.paginator.page.has_next(),
+                'has_previous': self.paginator.page.has_previous(),
+            }
+            return self.render_response(data, request, template_name=self.template_name, context=pagination_data)
+        
+        return self.render_response(data, request, template_name=self.template_name)
     
     def retrieve(self, request, *args, **kwargs):
         instance = self.get_object()
-        # Forzar respuesta en JSON si se pasa un parámetro en la URL ?format=json
-        if request.accepted_renderer.format == 'html':
-            return Response({'perro': instance}, template_name='perrito_detail.html')
-        serializer = self.get_serializer(instance)
-        return Response(serializer.data)
+        data = self.get_serializer(instance).data
+        return self.render_response(data, request, template_name='perrito_detail.html')
 
 
-#Form para cargar una nueva instancia
-class PerritosFormView(FormView):
+#View con form para cargar una nueva instancia
+class PerritosFormView(LoginRequiredMixin, FormView):
     form_class = PerrosForm
     template_name = 'perrito_create.html'
     success_url = reverse_lazy('adopta-list')
+    #login_url = "/login/"
+    #redirect_field_name = "redirect_to"
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -59,6 +84,10 @@ class PerritosFormView(FormView):
 
     def form_valid(self, form):
         files = self.request.FILES.getlist('file_field')
+        if not files:
+            form.add_error(None, 'Debe subie al menos una imagen.')
+            return self.form_invalid(form)
+        
         perro = form.save(commit = False)
         perro.user = self.request.user
         perro.save()
@@ -73,12 +102,14 @@ class PerritosFormView(FormView):
         return self.render_to_response(self.get_context_data(form=form))
 
 
-#View para editar las entradas
-class PerritosUpdateView(UpdateView):
+#View con form para editar las entradas
+class PerritosUpdateView(LoginRequiredMixin, UpdateView):
     model = Perro
     form_class = PerrosForm
     template_name = 'perrito_update.html'
     success_url = reverse_lazy('adopta-list')
+    #login_url = "/login/"
+    #redirect_field_name = "redirect_to"
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -89,8 +120,12 @@ class PerritosUpdateView(UpdateView):
     def form_valid(self, form):
         perro = form.save(commit=False)
         perro.save()
-        PerroFotos.objects.filter(perro=perro).delete()
         files = self.request.FILES.getlist('file_field')
+        if not files:
+            form.add_error(None, 'Debe subir al menos una imagen.')
+            return self.form_invalid(form)
+
+        PerroFotos.objects.filter(perro=perro).delete()
         for f in files:
             PerroFotos.objects.create(perro=perro, imagen=f)
